% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_helpers.R
\name{address_to_well_id}
\alias{address_to_well_id}
\title{Retrieve a well_id from barcode and address strings}
\usage{
address_to_well_id(
  barcode,
  address,
  .db_loc = "./htCE.duckdb",
  .db_con = NULL,
  .pg_load = FALSE
)
}
\arguments{
\item{barcode}{A string - the unique barcode identifying the physical plate}

\item{address}{A string - the well identifier (e.g. A1, H12, etc.)}

\item{.db_loc}{A string - local path to the DuckDB file; passed to \code{\link[=dkdb_collect]{dkdb_collect()}}}

\item{.db_con}{A valid DBIConnection object; passed to \code{\link[=dkdb_collect]{dkdb_collect()}}}

\item{.pg_load}{A Boolean - use the DuckDB Postgres extension?; passed to \code{\link[=dkdb_collect]{dkdb_collect()}}}
}
\value{
A double value representing the \code{well_id} integer primary key which corresponds to the barcode and address arguments
}
\description{
Retrieve a well_id from barcode and address strings
}
