% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_helpers.R
\name{barcode_to_plate_id}
\alias{barcode_to_plate_id}
\title{Retrieve a plate_id from a barcode string}
\usage{
barcode_to_plate_id(
  barcode,
  .db_loc = "./htCE.duckdb",
  .db_con = NULL,
  .pg_load = FALSE
)
}
\arguments{
\item{barcode}{A string - the unique barcode identifying the physical plate}

\item{.db_loc}{A string - local path to the DuckDB file; passed to \code{\link[=dkdb_collect]{dkdb_collect()}}}

\item{.db_con}{A valid DBIConnection object; passed to \code{\link[=dkdb_collect]{dkdb_collect()}}}

\item{.pg_load}{A Boolean - use the DuckDB Postgres extension?; passed to \code{\link[=dkdb_collect]{dkdb_collect()}}}
}
\value{
A double value representing the \code{plate_id} integer primary key which corresponds to the barcode argument
}
\description{
Retrieve a plate_id from a barcode string
}
