% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_helpers.R
\name{dkdb_lookup}
\alias{dkdb_lookup}
\title{Lookup something by a set of attributes and return the \code{id} or a Boolean value indicating if it already exists}
\usage{
dkdb_lookup(
  ...,
  .table = NULL,
  .return = "id",
  .db_loc = "./htCE.duckdb",
  .db_con = NULL,
  .pg_load = FALSE
)
}
\arguments{
\item{...}{Key-value pairs - unique set of injected arguments (using \code{\link[glue:glue_sql]{glue::glue_sql()}}) for component lookup}

\item{.table}{A string - the table to query in the database schema}

\item{.return}{A string - should the returned value be a \code{component_id} ("id") or a Boolean from an \code{EXISTS} operator subquery ("exists")}

\item{.db_loc}{A string - local path to the DuckDB file; passed to \code{\link[=dkdb_collect]{dkdb_collect()}}}

\item{.db_con}{A valid DBIConnection object; passed to \code{\link[=dkdb_collect]{dkdb_collect()}}}

\item{.pg_load}{A Boolean - use the DuckDB Postgres extension?; passed to \code{\link[=dkdb_collect]{dkdb_collect()}}}
}
\value{
A primary key \code{id} as a double (.return = "id") or a Boolean value returned from an \code{EXISTS} operator subquery (.return = "exixts")
}
\description{
Lookup something by a set of attributes and return the \code{id} or a Boolean value indicating if it already exists
}
