% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_helpers.R
\name{existing_barcodes}
\alias{existing_barcodes}
\title{Title}
\usage{
existing_barcodes(
  source_only = FALSE,
  .db_loc = "./htCE.duckdb",
  .db_con = NULL
)
}
\arguments{
\item{source_only}{A Boolean - only retrieve source plate barcodes}

\item{.db_loc}{A string - local path to the DuckDB file; passed to \code{\link[=dkdb_collect]{dkdb_collect()}}}

\item{.db_con}{A valid DBIConnection object; passed to \code{\link[=dkdb_collect]{dkdb_collect()}}}
}
\value{
A named character vector of barcodes that exist in the database
}
\description{
Title
}
