% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{count_newlines}
\alias{count_newlines}
\title{Return the number of lines in a file}
\usage{
count_newlines(path, remove_header = TRUE, ignore_trailing = TRUE)
}
\arguments{
\item{path}{a string - path to file}

\item{remove_header}{a Boolean - should the first row be ignored?}

\item{ignore_trailing}{a Boolean - should a trailing '\n' be ignored?}
}
\description{
Return the number of lines in a file
}
