% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_query.R
\name{db_query}
\alias{db_query}
\alias{dkdb_execute}
\alias{dkdb_collect}
\title{DuckDB Query}
\usage{
dkdb_execute(
  query_string,
  ...,
  .quiet = TRUE,
  .db_loc = "./htCE.duckdb",
  .db_con = NULL,
  .pg_install = FALSE,
  .pg_load = FALSE
)

dkdb_collect(
  query_string,
  ...,
  .quiet = TRUE,
  .db_loc = "./htCE.duckdb",
  .db_con = NULL,
  .pg_install = FALSE,
  .pg_load = FALSE
)
}
\arguments{
\item{query_string}{A string - SQL query}

\item{...}{Key-value pairs - additional injected values passed to \code{\link[glue:glue_sql]{glue::glue_sql()}}}

\item{.quiet}{A Boolean - supress console messages?}

\item{.db_loc}{A string - local path to the DuckDB file}

\item{.db_con}{A valid DBIConnection object}

\item{.pg_install}{A Boolean - install the DuckDB Postgres extension?}

\item{.pg_load}{A Boolean - use the DuckDB Postgres extension?}
}
\value{
A \code{tibble} containing collected results.
}
\description{
Execute or collect results from a query to a DuckDB database.
}
\details{
\code{dkdb_execute} executes a query without collecting results.

\code{dkdb_collect} collects the results from a query and returns them as a \code{tibble}.
}
