% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htce-package.R
\docType{package}
\name{htce-package}
\alias{htce}
\alias{htce-package}
\title{htce: A set of internal tools for managing high-throughput assay data at NEB}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Eric Hunt \email{52219938+eric-hunt@users.noreply.github.com} (\href{https://orcid.org/0000-0002-6891-8294}{ORCID})

}
\keyword{internal}
