% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_submission_files}
\alias{make_submission_files}
\title{Create CE submission files}
\usage{
make_submission_files(
  initials = "EH",
  datetime = NULL,
  plate_code = 5,
  num_wells = 96,
  num_plates = 12,
  dest_dir = NULL,
  .pad_well_num = FALSE
)
}
\arguments{
\item{initials}{A string - your initials}

\item{datetime}{A date - use \code{base::as.Date} to convert a string to a date}

\item{plate_code}{An integer - a code representing the type of plate}

\item{num_wells}{An integer - 96 or 384, the number of wells on the submission plate}

\item{num_plates}{An integer - the number of plate submission files to create}

\item{dest_dir}{A string - the directory to create the tab-delimited submission files}

\item{.pad_well_num}{A Boolean - passed to \code{\link[=make_well_order]{make_well_order()}} \emph{pad} argument}
}
\description{
Creates tab-delimited files for submitting CE plates to the NEB sequencing core via the LIMS system.
}
