% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_well_order}
\alias{make_well_order}
\title{Create a vector of well addresses}
\usage{
make_well_order(num_wells = 96L, by = c("column", "row"), pad = FALSE)
}
\arguments{
\item{num_wells}{An integer - the number of wells in the plate format (96 or 384)}

\item{by}{A string - which way the order should be populated relative to the plate format, i.e. by \code{"column"} (default) or by \code{"row"}}

\item{pad}{A Boolelan - whether or not to pad the column numbers with zeros (e.g. A1 vs. A01)}
}
\value{
A character vector of well addresses in top to bottom, left to right order (e.g. for 96 wells, A1 -> H1, A2...H11, A12 -> H12)
}
\description{
Create a vector of well addresses
}
