% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_connection.R
\name{with_duckdb_connection}
\alias{with_duckdb_connection}
\title{DuckDB Connection}
\usage{
with_duckdb_connection(
  db_file_path,
  read_only = FALSE,
  .local_envir = parent.frame()
)
}
\arguments{
\item{db_file_path}{a string - path to a DuckDB file}

\item{read_only}{a Boolean - passed to \code{\link[DBI:dbConnect]{DBI::dbConnect()}};
default is \code{FALSE}}

\item{.local_envir}{an environment - passed to \code{\link[withr:defer]{withr::defer()}};
default is \code{parent.frame()}}
}
\value{
A database connection of class \emph{DBIConnection}.
}
\description{
DuckDB connection handling à la \code{withr::local_db_connection}.
}
\seealso{
\code{\link[DBI:dbConnect]{DBI::dbConnect()}}, \code{\link[withr:defer]{withr::defer()}}, \code{\link[withr:local_db_connection]{withr::local_db_connection()}},
\href{https://github.com/duckdb/duckdb/pull/5525}{duckdb/duckdb#5525}
}
